Remove BrowserBasedDashboard Error E2E Test
Remvoing the BrowserBasedDashboard_ErrorThrownMissingMandatorySection due to upcoming changes on error handling on non-form sections (this page is going to be changed to never be able to throw an error).

Fix ErrorSummaryLinksExist to actually extract the ElementId from the link using an extra select and change to an ElementExists check instead of ElementIsDisplayed since the displayed logic may be held on a parent tag.