Add Edit Solution and Existing Solution E2E Tests
Add Delete Solution Tests

Update Order Session State Model to have "HasHitEditSolution" to mark if we have been to the edit solution page. this is so that we can redirect correctly from the select recipients page to the correct target depending on where the logic journey came from.

Update InitialiseStateForEdit to set the above value when the Edit solution page is loaded.

Remove unneeded Usings.