Add Cat Solutions Capabilities and Epics Filter
Add Catalogue Solutions and Capabilites Filter to the Catalogue
Solutions Page.

Update CapabilitiesCategories Table Schema to Include Description

Update Capabilities to include their correct category Id

Add CatalogueSolutionsFilter JS that will reload the the filter Expander
when javascript is turned on for the client and add in the requried
logic to make the capbilities update when you select different frameworks

Update the Pagination Logic to include existing Query Parameters when
it adds the page parameter.

add ISolutionFilterService and logic to do all the hard work of
running the filter.

Add the PaginationBuilder to ISolutionFilterService to dynamically
build the where clause depending on the selected Capabilities
and epics.

Add cache logic to the new "Filter" action in SolutionsController
and into two endpoints in ISolutionsFilterService to cache the HTML
response for the different filter types since these won't update
very often and are quite costly to render.

Add new CacheKey to settings so that in the future the caches
will only decache when a solution is changed. right now it just
defaults to 60 second lifetime.