Add Supplier Management Pages
Move all Supplier Management Pages into Suppliers Controller

Add new SuppliersService for managing handling of the details

Add Models and View for Handling Supplier Managment

Update FieldSetContainerTagHelper and FieldSetFormLableTagHelper to correctly not set that it can handle a field counter.

Update UpdateOutputDiv to only apply a surrounding div to a form-group if there is a counter on the field
if there is no counter, then it will set the form-group as the top level by passing on all its attributes
and then removing itself. (this is so that inputs can work with fieldset-containers).

Move supplier related logic out of SolutionsSerivce

Move Supplier related controllers out of HomeController

Add SupplierManagement.js to handle the showing and hiding of inactive suppliers on the manage all suppliers page.

Update existing tests to reflect the new arrangement Supplier logic moving from solutionsSerive to SuppliersSerivce