Update Solution Editing Dashboard
Update Solution Editing Dashboard with Feedback from Kirsty and Miranda

Make the Dashboard a two-thirds column instead of a full width due to the fact the edit button is out of your eyeline when reading the section name.

add a preview button that takes you to the marketing pages, so that you can see how a solution will look in marketing before you publish it.

reorder the sections to follow the section order of the solution PDF that is handed to kirsty for uploading.