Requests From Kirsty
Update EditCapabilities.cshtml to show capability and epic Ref/Id before the name so that they're easier to find in the list.

Update EditWorkOffPlanValidator.cs to allow work-off plans 2 months in the past.

Upate CapabilitiesService.GetCapabilitiesByCategory to only include supplier defined epics of the solution you are looking ats supplier.