Fix additional and Associated Service Marketing Pages
Update the views, models, controller and clean up the SolutionService so it doesn't bring back a huge amoutn of duff data.

Update how additional and associated services get their data.

Fix Edit Assoc/Addit details pages to not throw errors for duplicate names when editing an item and itself is the only one with that name.

Fix tests