Fix Select Contact on Solution Marketing Page Being Incorrect
Fix the bug on the solutions marketing page where the selected supplier contact was not updating when changed.

The reason for this was that the supplier contacts save into the supplier contacts are saved into the supplier contacts table, but the marketing page was using the marketing contacts table.

Updated Marketing page to get marketing contats from CatalogueItemContacts Table Reference.

Update GetSolutionWithSupplierDetails to not bring back the marketing contacts.

Fix Tests and Update Seed Data