Fix form errors on add a user page
Update AddUserModelValidator to cover all form error scenarios

Update CreateBuyerService to no longer handle the form validation

Delete the old CreateBuyerService validation logic and tests

Delete the Results object since it's no longer being used

Add AddUserModelValidator Tests

Split AddUser E2E Tests into their own class

Update Test base to handle a bad logic scenario.
when a test logs you in, the next test will also see if you need to log in by checking for the existance of an input with the title "emailAddress". if the page the next test loads has an input called "emailAddress" it will then try to log in on that form and fail.
Change the Login functions to check if the user is already logged in by checking for the user-session cookie.