Update CatalogueItemIdValueGenerator.cs
Update CatalogueItemIdValueGenerator to correctly scope the ID of an additional service to the solution that it belongs to.

Currently, in live, the query searches all additional services belonging to a supplier resulting in cross-contamination of additional service IDs. The correct behavior is to relate the additional service IDs to the solution.